% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_null_models.R
\name{coef_plot}
\alias{coef_plot}
\title{Model coefficients inference plots based on a given set of null model objects}
\usage{
coef_plot(model, null_model_sums, coefs = NULL)
}
\arguments{
\item{model}{A fitted model object. Currently, objects of classes 'lm', 'glm', and 'gam' are supported.}

\item{coefs}{An optional character vector specifying which model coefficients should be inferred. If not specified,
the inference is made for all model coefficients.}

\item{null_models}{A list of model objects of the same class as model representing simulations of a null model.}
}
\value{
A ggplot with a facet for each model coefficient, plotting its observed value (red line) on a density of
its null distribution as well as 0.025, 0.5, and 0.975 quantiles.
}
\description{
This function plots the observed value of each model coefficient and compares it with a null distribution of this
coefficient based on a set of null model objects. These null models are typically created by the function
simulate_null_models from the same package. The plots are based on (and thus require) ggplot2 package.
}
