% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_p_values.R
\name{drop_one}
\alias{drop_one}
\title{Significance of predictors based on simulated null models}
\usage{
drop_one(model, null_models, preds = NULL)
}
\arguments{
\item{model}{A fitted model object. Currently, objects of classes 'lm', 'glm', and 'gam' are supported.}

\item{null_models}{A list of model objects of the same class as model representing simulations of a null model.}

\item{preds}{An optional character vector specifying for which model coefficients the inference should be made.
If not specified,the inference is made for all coefficients.}
}
\value{
A data frame summarizing the null distribution for each model coefficient, including a Monte Carlo p-value.
}
\description{
This function summarizes the null distribution of predictor p values based on a given set of null model objects
(typically created by the function simulate_null_models from teh same package) and compares the observed p value
with this null distribution.
}
