% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_null_models.R
\name{effect_plot}
\alias{effect_plot}
\title{Plotting the effects of individual continuous predictors}
\usage{
effect_plot(
  model,
  null_models,
  data,
  preds = NULL,
  lower = c(0.025),
  upper = c(0.975),
  nval = 100,
  plotdata = TRUE
)
}
\arguments{
\item{model}{A fitted model object. Currently, objects of classes 'lm', 'glm', and 'gam' are supported.}

\item{null_models}{A list of model objects of the same class as model representing simulations of a null model.}

\item{data}{A data frame used to fit the model.}

\item{preds}{An optional character vector of names of predictors to be plotted. If not specified, all continuous
predictors used to fit the model are used.}

\item{lower}{The lower quantile of the simulation envelope to be plotted. Default is 0.025.}

\item{upper}{The upper quantile of the simulation envelope to be plotted. Default is 0.975.}

\item{nval}{The number of values (between minimum and maximum from its observed values) of each predictor to be used
for plotting the effect. Default is 100.}

\item{plotdata}{A logical indicating whether the data points should be plotted together with the effect. Default is TRUE.}
}
\value{
A ggplot object with a facet for each model predictor, plotting the predicted values (red line) together with
simulation envelopes (grey shaded area) and median (black line).
}
\description{
This function plots the effects of individual continuous predictors, keeping the values of other predictors at their
mean values, and showing 95\% envelope based on given null model simulations. It can be used for a visual check whether
the observed effect could be explained by the null model or not. Typically, the null model relaxes any structural
relationships between the response and the predictors while keeping the autocorrelation structure unchanged.
}
