% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_null_models.R
\name{prepare_data}
\alias{prepare_data}
\title{Function to prepare data for the effect plot, preparing the sequence of values of the predictor pred
while keeping the values of other continuous predictors at their mean value.}
\usage{
prepare_data(data, pred, nval)
}
\description{
Function to prepare data for the effect plot, preparing the sequence of values of the predictor pred
while keeping the values of other continuous predictors at their mean value.
}
