% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_null_models.R
\name{simulate_null_models}
\alias{simulate_null_models}
\title{Simulate null models}
\usage{
simulate_null_models(
  model,
  data,
  preds = NULL,
  pred_ras = NULL,
  variog = NULL,
  coords = c("x", "y"),
  method = c("shift", "RFsim", "Viladomat", "kriging", "shift_only", "rotate_only"),
  output = c("coef", "p", "dev", "AIC", "AUC", "R2", "MSE"),
  formulas = NULL,
  radius = NULL,
  nsim = 1000
)
}
\arguments{
\item{model}{A fitted model object. Currently, objects of classes 'lm', 'glm', 'gam', and 'ranger' are supported.}

\item{data}{A data frame with the original dataset used to fit the model, including spatial coordinates of the data points.}

\item{preds}{A character vector of names of predictors, for which the null models should be generated. If NULL, all predictors
from the model are used (except coordinates themselves, in case they were used as predictors). Note that for some methods,
corresponding predictor rasters must be passed through the argument pred_ras (see Details).}

\item{pred_ras}{A named list of predictor rasters, with one raster for each predictor. (Only one predictor is currently
supported). The rasters in the list must be named the same as the original predictors. The argument is optional, only
required when method is one of 'shift', 'shift_only', and 'rotate only'.}

\item{variog}{A variogram object of the class 'variogram' from the package geoR (Ribeiro et al. 2020). An optional argument,
ignored unless the method is 'RFsim'. See Details.}

\item{coords}{A character vector of lenght two with names of the columns in data where spatial coordinates are stored.
Defaults are 'x' and 'y'.}

\item{method}{A method how the null models are generated. Possible values are 'shift','RFsim','Viladomat','kriging',
'shift_only',and 'rotate_only'. See Details for their description. Default is 'shift'.}

\item{output}{There will be an explanation here...}

\item{formulas}{There will be an explanation here...}

\item{radius}{A maximum distance for random shift of the data points. Only used when method is 'shift' or 'shift_only'
(see Details).}

\item{nsim}{The number of simulated null models to return.}
}
\value{
A list of summaries of simulated models. The first item correspond to the original model, the rest correpond to the
models fitted to simulated data. The length of the list is nsim + 1.

Here, the description of the individual list items will be included.
}
\description{
This is the main function, simulating a set of null models based on a given model and data, keeping the original
spatial autocorrelation structure unchanged, but relaxing any original structural relationship between the response
and predictors.
}
\details{
The method 'shift' randomly move the data points (all in the same way) by first rotating the dataset around the data
centroid by a random angle, and then shifting them by in a random direction by a distance generated as a uniform random
variable on a (0, radius) distance. Then, new predictor values are extracted at these new locations from the corresponding
predictor raster passed through the argument pred_ras. Methods 'shift_only' and 'rotate_only' perform only the corresponding
part of the procedure (i.e., either random shift or random rotation).

The method 'RFsim' generate the new datasets by unconditional simulation of random fields with given variogram. the simulation
is performed using the function RFsimulate from the package RandomFields (Schlather et al. 2022). The variogram is either
known and passed through the variog argument, or unknown and estimated from data using maximum likelihood method (function
likfit from the geoR (Ribeiro et al. 2020) package).

The method 'kriging' is based on first random rotation and shift of the data points (same as for the method 'shift'), and then
estimation of the new values at these locations by performing kriging. Similarly to the method RFsim, if the argument variog is
passed (i.e., is not NULL), this variogram is used for the kriging. If variog is NULL, the variogram needed for the kriging
is estimated from data by maximum likelihood method.

The method 'Viladomat' is based on random permutation of the predictor values and then smoothing the values so that their
variogram approximately matches the original variogram. the method is described in Viladomat et al. 2014.
}
